I've restored "Delete" button from each table's row into SubHeaderComponent.
How to remove necessary rows?
Step 1: Click on "Select a row" button to open rows' checkboxes.
Step 2: Click on checkboxes of lines that should be removed.
Step 3: Click on "Delete N row(s)" button to open a confirm dialog. If you dont want to remove that selected rows, click on "No" (and you can choose other rows or close the table); if necessary, click on "Yes" (and selected rows will be removed from the table).