It is impossible to import customized TableComponent on the server-side rendering page.
So I've created a middle file for Users Data Table between SSR page and customized for different data objects types TableComponent. I think it is the excellent idea.

This Users Data Table has already been getting data from sqlite3 database. The next step will be adding functions to save changes and new rows on the DB through server-side API.

And, of course, TableComponent will be divided on parts for ease (relief) settings them in it and for another app components.