3 language localization has been added for (admin) route group.
When a unique authentication code is generated, it won't include next letters: "I" (it looks like 1 (one)) and "O" (it is simular as 0 (zero)).

Admin shouldn't click on "Shift" and "-" (hyphen) buttons while inputting a unique generated code from his/her email. Letters are written in uppercase and a hyphen is added after the fourth and 10th symbol of unique generated code. (If you will need to delete it, select all symbols at once and click on "Delete" keyboard button).