The first SortableItem component is named "Ўпадабаныя Стравы" ("Favourite Dishes") and includes 3 favourite dishes for John Doe (using UsersDishes table for many-to-many relationship between "Users" and "dishes_BY" tables).
Each dish is shown as Card component with a dish image on all Card's square, dish title in the bottom and 2 buttons on the top-right.

The first one will be for getting the common information about a clicked dish as an intercepting route.
The second one will be for removing one or more choisen dishes from "Favourite Dishes" block.