All pages for an authorized user have started supporting page styles and light/dark themes.
Admin's Roles Table has got its own Expander component and has started working well.

New additional information is shown in the bottom side of the page of a choisen dish: when a dish is usually eaten: for breakfast, for lunch, and/or for dinner. Of course, it supports all available language localizations of the app.