The special context has been added for connection between 2 client components: MenuButton and Navbar – to show and hide the last one by clicking on the first.
I have already done home cafe manager page and, unfortunately, I have no plan what to do next. The only thing I'm thinking about is the implementation of geomap for showing cafe on it. Every cafe has its own geo-data in database: latitude, longitude and SRId.

So I will be looking for this geomap tomorrow and will add it for the app.