Actual Dishes part of "cooking&serving" page has got a choisen dishes panel at the bottom of a bowser's window named BuyingLane.
You can increment/decrement portions of an each choisen dish in the range between 1 and 3 inclusively. Also if you want to remove a dish from the panel, you can do it cliking on the trash icon on the dish image. Its focus will be removed from actual dishes' categories too.

The last main step for this part is adding the payment and preview window before the payment.